backend/requirements/development.txt
-r ./base.txt

django-extensions==3.1.5

pytest==7.1.2
pytest-cov==3.0.0
pytest-django==4.5.1
pytest-mock==3.7.0

Faker==13.14.0
ipdb==0.13.9
factory_boy==3.2.1
freezegun==1.2.1

Sphinx==5.0.2
djangorestframework==3.13.1
django-allauth==0.43.0
django-rest-auth==0.9.5
